Added BottomNavigationBar and extracted PokemonListPage from HomePage body
