+++ lib/common/widgets/common_scaffold.dart
@@ -1,15 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:pokedex_rest/common/widgets/common_bottom_navigation_bar.dart';
 import 'package:pokedex_rest/style/color_tokens.dart';
 
 class CommonScaffold extends StatelessWidget {
   final Widget body;
-  final PreferredSizeWidget appBar;
-  final CommonBottomNavigationBar? navigationBar;
+  final PreferredSizeWidget? appBar;
+  final BottomNavigationBar? navigationBar;
 
   const CommonScaffold({
     required this.body,
-    required this.appBar,
+    this.appBar,
     this.navigationBar,
     Key? key,
   }) : super(key: key);
