+++ lib/style/dimensions.dart
@@ -2,6 +2,7 @@ class Dimensions {
   static const double size0 = 0.0;
   static const double sizeS = 2.0;
   static const double sizeM = 6.0;
+  static const double sizeML = 8.0;
   static const double sizeL = 10.0;
   static const double sizeLXL = 12.0;
   static const double sizeXL = 16.0;
